Zombie Spawn Logic
Zombies now move to a goal location when spawned, Using the Spawn function, they will move to this goal and then move to the player. This allows zombies to move towards their nearest door regardless of the player location.

When placing spawnpoints there is an info script to be included that allows you to allocate a goal location, all set up in the spawning backend.